+++ client/src/styles/Home.css
@@ -1,4 +1,4 @@
-.app-header {
+.home-header {
   background-color: #282c34;
   min-height: 100vh;
   display: flex;
